refactor app.js by including routes
